+++ android/src/main/java/com/bitmovin/player/reactnative/PlayerModule.kt
@@ -50,13 +50,31 @@ class PlayerModule(private val context: ReactApplicationContext) : ReactContextB
     /**
      * Load the source of the given `nativeId` with `config` options from JS.
      * @param nativeId Target player.
-     * @param config Source configuration options from JS.
+     * @param sourceNativeId Target source.
      */
     @ReactMethod
     fun loadSource(nativeId: NativeId, sourceNativeId: String) {
         uiManager()?.addUIBlock {
             sourceModule()?.getSource(sourceNativeId)?.let {
                 players[nativeId]?.load(it)
+            }
+        }
+    }
+
+    /**
+     * Load the `offlineSourceConfig` for the player with `nativeId` and offline source module with `offlineModuleNativeId`.
+     * @param nativeId Target player.
+     * @param offlineContentManagerBridgeId Target offline module.
+     * @param options Source configuration options from JS.
+     */
+    @ReactMethod
+    fun loadOfflineContent(nativeId: NativeId, offlineContentManagerBridgeId: String, options: ReadableMap?) {
+        uiManager()?.addUIBlock {
+            val offlineSourceConfig = offlineModule()?.getOfflineContentManagerBridge(offlineContentManagerBridgeId)
+                ?.offlineContentManager?.offlineSourceConfig
+
+            if (offlineSourceConfig != null) {
+                players[nativeId]?.load(offlineSourceConfig)
             }
         }
     }
@@ -399,6 +417,18 @@ class PlayerModule(private val context: ReactApplicationContext) : ReactContextB
     }
 
     /**
+     * Sets the max selectable bitrate for the player.
+     * @param nativeId Target player id.
+     * @param maxSelectableBitrate The desired max bitrate limit.
+     */
+    @ReactMethod
+    fun setMaxSelectableBitrate(nativeId: NativeId, maxSelectableBitrate: Int) {
+        uiManager()?.addUIBlock {
+            players[nativeId]?.setMaxSelectableVideoBitrate(maxSelectableBitrate.takeUnless { it == -1 } ?: Integer.MAX_VALUE)
+        }
+    }
+
+    /**
      * Helper function that returns the initialized `UIManager` instance.
      */
     private fun uiManager(): UIManagerModule? =
@@ -409,4 +439,10 @@ class PlayerModule(private val context: ReactApplicationContext) : ReactContextB
      */
     private fun sourceModule(): SourceModule? =
         context.getNativeModule(SourceModule::class.java)
+
+    /**
+     * Helper function that returns the initialized `OfflineModule` instance.
+     */
+    private fun offlineModule(): OfflineModule? =
+        context.getNativeModule(OfflineModule::class.java)
 }