+++ android/src/main/java/com/bitmovin/player/reactnative/converter/JsonConverter.kt
@@ -1,36 +1,52 @@
 package com.bitmovin.player.reactnative.converter
 
 import com.bitmovin.analytics.BitmovinAnalyticsConfig
-import com.bitmovin.analytics.config.SourceMetadata
-import com.bitmovin.analytics.data.CustomData
+import com.bitmovin.analytics.api.CustomData
+import com.bitmovin.analytics.api.SourceMetadata
 import com.bitmovin.player.api.DeviceDescription.DeviceName
 import com.bitmovin.player.api.DeviceDescription.ModelName
 import com.bitmovin.player.api.PlaybackConfig
 import com.bitmovin.player.api.PlayerConfig
 import com.bitmovin.player.api.TweaksConfig
-import com.bitmovin.player.api.advertising.*
+import com.bitmovin.player.api.advertising.Ad
+import com.bitmovin.player.api.advertising.AdBreak
+import com.bitmovin.player.api.advertising.AdConfig
+import com.bitmovin.player.api.advertising.AdData
+import com.bitmovin.player.api.advertising.AdItem
+import com.bitmovin.player.api.advertising.AdQuartile
+import com.bitmovin.player.api.advertising.AdSource
+import com.bitmovin.player.api.advertising.AdSourceType
+import com.bitmovin.player.api.advertising.AdvertisingConfig
 import com.bitmovin.player.api.drm.WidevineConfig
 import com.bitmovin.player.api.event.PlayerEvent
 import com.bitmovin.player.api.event.SourceEvent
 import com.bitmovin.player.api.event.data.SeekPosition
+import com.bitmovin.player.api.media.AdaptationConfig
 import com.bitmovin.player.api.media.audio.AudioTrack
 import com.bitmovin.player.api.media.subtitle.SubtitleTrack
 import com.bitmovin.player.api.media.thumbnail.ThumbnailTrack
 import com.bitmovin.player.api.media.video.quality.VideoQuality
+import com.bitmovin.player.api.offline.options.OfflineContentOptions
+import com.bitmovin.player.api.offline.options.OfflineOptionEntry
 import com.bitmovin.player.api.source.Source
 import com.bitmovin.player.api.source.SourceConfig
+import com.bitmovin.player.api.source.SourceOptions
 import com.bitmovin.player.api.source.SourceType
+import com.bitmovin.player.api.source.TimelineReferencePoint
 import com.bitmovin.player.api.ui.ScalingMode
 import com.bitmovin.player.api.ui.StyleConfig
 import com.bitmovin.player.reactnative.extensions.getName
-import com.bitmovin.player.reactnative.extensions.putInt
+import com.bitmovin.player.reactnative.extensions.getProperty
 import com.bitmovin.player.reactnative.extensions.putDouble
+import com.bitmovin.player.reactnative.extensions.putInt
+import com.bitmovin.player.reactnative.extensions.setProperty
 import com.bitmovin.player.reactnative.extensions.toList
 import com.bitmovin.player.reactnative.extensions.toReadableArray
-import com.bitmovin.player.reactnative.extensions.getProperty
-import com.bitmovin.player.reactnative.extensions.setProperty
-import com.facebook.react.bridge.*
 import com.bitmovin.player.reactnative.extensions.toReadableMap
+import com.facebook.react.bridge.Arguments
+import com.facebook.react.bridge.ReadableArray
+import com.facebook.react.bridge.ReadableMap
+import com.facebook.react.bridge.WritableMap
 import java.util.UUID
 
 /**
@@ -71,7 +87,51 @@ class JsonConverter {
                     playerConfig.advertisingConfig = it
                 }
             }
+            if (json.hasKey("adaptationConfig")) {
+                toAdaptationConfig(json.getMap("adaptationConfig"))?.let {
+                    playerConfig.adaptationConfig = it
+                }
+            }
             return playerConfig
+        }
+
+        /**
+         * Converts an arbitrary `json` to `SourceOptions`.
+         * @param json JS object representing the `SourceOptions`.
+         * @return The generated `SourceOptions`.
+         */
+        @JvmStatic
+        fun toSourceOptions(json: ReadableMap?): SourceOptions {
+            if (json == null) return SourceOptions()
+            val startOffset = if(json.hasKey("startOffset")) json.getDouble("startOffset") else null
+            val timelineReferencePoint = toTimelineReferencePoint(json.getString("startOffsetTimelineReference"))
+            return SourceOptions(startOffset = startOffset, startOffsetTimelineReference = timelineReferencePoint)
+        }
+
+        /**
+         * Converts an arbitrary `json` to `TimelineReferencePoint`.
+         * @param json JS string representing the `TimelineReferencePoint`.
+         * @return The generated `TimelineReferencePoint`.
+         */
+        @JvmStatic
+        private fun toTimelineReferencePoint(json: String?): TimelineReferencePoint? = when (json) {
+            "start" -> TimelineReferencePoint.Start
+            "end" -> TimelineReferencePoint.End
+            else -> null
+        }
+        
+        /**
+         * Converts an arbitrary `json` to `AdaptationConfig`.
+         * @param json JS object representing the `AdaptationConfig`.
+         * @return The generated `AdaptationConfig` if successful, `null` otherwise.
+         */
+        private fun toAdaptationConfig(json: ReadableMap?): AdaptationConfig? {
+            if (json == null) return null
+            val adaptationConfig = AdaptationConfig()
+            if (json.hasKey("maxSelectableBitrate")) {
+                adaptationConfig.maxSelectableVideoBitrate = json.getInt("maxSelectableBitrate")
+            }
+            return adaptationConfig
         }
 
         /**
@@ -213,7 +273,7 @@ class JsonConverter {
                 ?.mapNotNull(::toAdSource)
                 ?.toTypedArray()
                 ?: return null
-            return AdItem(sources, json?.getString("position") ?: "pre")
+            return AdItem(sources, json.getString("position") ?: "pre")
         }
 
         /**
@@ -271,6 +331,9 @@ class JsonConverter {
                 config.metadata = json.getMap("metadata")
                     ?.toHashMap()
                     ?.mapValues { entry -> entry.value as String }
+            }
+            if (json.hasKey("options")) {
+                config.options = toSourceOptions(json.getMap("options"))
             }
             return config
         }
@@ -545,7 +608,7 @@ class JsonConverter {
                 false
             }
             val format = json.getString("format")
-            if (format != null && format.isNotBlank()) {
+            if (!format.isNullOrBlank()) {
                 return SubtitleTrack(
                     url = url,
                     label = label,
@@ -778,17 +841,20 @@ class JsonConverter {
          * @return The produced `CustomData` or null.
          */
         @JvmStatic
-        fun toAnalyticsCustomData(json: ReadableMap?): CustomData? = json?.let {
-            val customData = CustomData()
-            for (n in 1..30) {
-                it.getString("customData${n}")?.let { customDataN ->
-                    customData.setProperty("customData${n}", customDataN)
+        fun toAnalyticsCustomData(json: ReadableMap?): CustomData? {
+            if (json == null) return null
+
+            return CustomData.Builder().apply {
+                for (n in 1..30) {
+                    setProperty(
+                        "customData${n}",
+                        json.getString("customData${n}") ?: continue
+                    )
                 }
-            }
-            it.getString("experimentName")?.let { experimentName ->
-                customData.experimentName = experimentName
-            }
-            customData
+                json.getString("experimentName")?.let {
+                    setExperimentName(it)
+                }
+            }.build()
         }
 
         /**
@@ -812,23 +878,15 @@ class JsonConverter {
 
         @JvmStatic
         fun toAnalyticsSourceMetadata(json: ReadableMap?): SourceMetadata? = json?.let {
-            val sourceMetadata = SourceMetadata(
-                    title = it.getString("title"),
-                    videoId = it.getString("videoId"),
-                    cdnProvider = it.getString("cdnProvider"),
-                    path = it.getString("path"),
-                    isLive = it.getBoolean("isLive")
+            val sourceCustomData = toAnalyticsCustomData(json) ?: CustomData()
+            SourceMetadata(
+                title = it.getString("title"),
+                videoId = it.getString("videoId"),
+                cdnProvider = it.getString("cdnProvider"),
+                path = it.getString("path"),
+                isLive = it.getBoolean("isLive"),
+                customData = sourceCustomData
             )
-
-            for (n in 1..30) {
-                it.getString("customData${n}")?.let { customDataN ->
-                    sourceMetadata.setProperty("customData${n}", customDataN)
-                }
-            }
-            it.getString("experimentName")?.let { experimentName ->
-                sourceMetadata.experimentName = experimentName
-            }
-            sourceMetadata
         }
 
         /**
@@ -846,6 +904,36 @@ class JsonConverter {
                 putDouble("frameRate", videoQuality.frameRate.toDouble())
                 putInt("height", videoQuality.height)
                 putInt("width", videoQuality.width)
+            }
+        }
+
+        /**
+         * Converts any `OfflineOptionEntry` into its json representation.
+         * @param offlineEntry `OfflineOptionEntry` object to be converted.
+         * @return The generated json map.
+         */
+        @JvmStatic
+        fun toJson(offlineEntry: OfflineOptionEntry): WritableMap {
+            return Arguments.createMap().apply {
+                putString("id", offlineEntry.id)
+                putString("language", offlineEntry.language)
+            }
+        }
+
+        /**
+         * Converts any `OfflineContentOptions` into its json representation.
+         * @param options `OfflineContentOptions` object to be converted.
+         * @return The generated json map.
+         */
+        @JvmStatic
+        fun toJson(options: OfflineContentOptions?): WritableMap? {
+            if (options == null) {
+                return null
+            }
+
+            return Arguments.createMap().apply {
+                putArray("audioOptions", options.audioOptions.map { toJson(it) }.toReadableArray())
+                putArray("textOptions", options.textOptions.map { toJson(it) }.toReadableArray())
             }
         }
     }